Changes in c# Assignment

+++ Daily Assignments/C#/Daily.Assg 4/Daily.Assg 4/Daily.Assg 4/IStudent.cs
@@ -10,7 +10,7 @@ namespace Daily.Assg_4
     {
         int StudentId { get; set; }
         string Name { get; set; }
-        double Fees { get; set; }
+        int Fees { get; set; }
         void ShowDetails();
     }
